0829 bullet 움직임 수정

+++ 0806_임시프로젝트파일_ 하영/GP_subproject_first/Build.h
@@ -18,6 +18,6 @@ public:
 
 	virtual void Rest(float Delta) override;
 
-	void CalcDirection(int xvec, int yvec) override;
+	void CalcDirection(float xvec, float yvec) override;
 	std::vector<Gdiplus::Rect> restRc;
 };
