+++ 0806_임시프로젝트파일_ 하영/GP_subproject_first/myUnit.h
@@ -31,7 +31,7 @@ public:
 	virtual void ExtraAction(float Delta);
 	virtual void CreateBT(BlackBoard* InBB);
 
-	virtual void CalcDirection(int xvec,int yvec);
+	virtual void CalcDirection(float xvec,float yvec);
 
 	//unit status
 	UnitInfo mUnitInfo;
